+++ src/terra-dev-site/test/modal-manager/DisclosureComponent.jsx
@@ -15,6 +15,7 @@ import {
 import NavigationPrompt from '../../../navigation-prompt';
 import styles from './DisclosureComponentCommon.test.module.scss';
 import ApplicationLoadingOverlay from '../../../application-loading-overlay';
+import ApplicationPageStatus from '../../../application-page-status';
 
 const cx = classNames.bind(styles);
 
@@ -44,6 +45,7 @@ class DisclosureComponent extends React.Component {
       hasPendingAction: false,
       hasLoadingOverlay: false,
       hasError: false,
+      hasPageStatus: false,
       disclosureCount: null,
     };
   }
@@ -143,7 +145,7 @@ class DisclosureComponent extends React.Component {
   render() {
     const { disclosureManager, identifier, renderHeaderAdapter } = this.props;
     const {
-      hasPendingAction, hasLoadingOverlay, hasError, disclosureCount,
+      hasPendingAction, hasLoadingOverlay, hasError, hasPageStatus, disclosureCount,
     } = this.state;
 
     if (hasError) {
@@ -183,6 +185,7 @@ class DisclosureComponent extends React.Component {
         <button type="button" className="disclose-fullscreen" onClick={this.disclose('fullscreen')}>Disclose Fullscreen</button>
         <button type="button" className="pending-action-toggle" onClick={() => { this.setState((state) => ({ hasPendingAction: !state.hasPendingAction })); }}>Pending Action</button>
         <button type="button" className="loading-overlay-toggle" onClick={() => { this.setState((state) => ({ hasLoadingOverlay: !state.hasLoadingOverlay })); }}>Loading Overlay</button>
+        <button type="button" className="status-view-toggle" onClick={() => { this.setState((state) => ({ hasPageStatus: !state.hasPageStatus })); }}>Status View</button>
         <button type="button" className="error-toggle" onClick={() => { this.setState((state) => ({ hasError: !state.hasError })); }}>Error</button>
         <div className={cx('content-form')}>
           {this.renderForm()}
@@ -196,6 +199,7 @@ class DisclosureComponent extends React.Component {
         <button type="button" className="global-close-disclosure" onClick={closeMostRecentDisclosure}>{`Global Close (${disclosureCount})`}</button>
         {hasPendingAction && <NavigationPrompt description="Test Action" />}
         {hasLoadingOverlay && <ApplicationLoadingOverlay isOpen backgroundStyle="light" />}
+        {hasPageStatus && <ApplicationPageStatus message="Test status view" variant="no-data" />}
       </ContentContainer>
     );
   }